Add alpine-golang image
This adds a build image based solely on golang:1.15-alpine.

Since with the new confluent-go-kafka version no actual installation of Kafka is required anymore, we can build apps in a golang-alpine container with nothing but go and alpine (and build tools). This image is smaller than the one with librdkafka and will thus download a bit smaller, resulting in faster pipelines.